feat: Move existing Skylark queries to classic folder
