+++ packages/ingestor/src/lib/skylark/classic/index.ts
@@ -2,3 +2,5 @@ export * from "./api";
 export * from "./create";
 export * from "./get";
 export * from "./sets";
+export * from "./availability";
+export * from "./content-types";
